requirements.txt
Flask~=2.3.2
nomic~=1.1.6
torch~=2.0.1
openai~=0.27.7
python-dotenv~=1.0.0
gTTS~=2.3.2
pyttsx3~=2.90
pydub~=0.25.1
deep_translator